64 updates to filingperiod and filing objects (#65)
Closes #64 

This is a decent size refactor so I wanted to get this out there. Once
https://github.com/cfpb/sbl-filing-api/pull/63 is approved I'll need to
merge it into here and redo a couple of things that were added in that
story (query_helper updates, etc)

I've put comments on certain changes defined below that I think will
warrant conversation. This is was all based on going with the path
/v1/filing/{lei}/{period_name}/ to get filing objects, post a new filing
object, and then later submissions would be under
/v1/filing/{lei}/{period_name}/submissions. Using LEI and period name
(referred to as filing_period in the child tables) as a composite
primary and foreign key drove the changes here.