Added 409 status for existing Filing on POST (#94)
Closes #82 

- On "empty" filing POST, check if the lei + period filing exists and if
so, throw a 409 CONFLICT result with message
- Updated filing post pytest

Let in the check if a FilingDTO is passed in because it sounds like
there may be a desire for the frontend to post the initial filing json.
However, this will require us making the FilingDTO id optional which may
not be desired. If it's determined the empty POST is fine, then we can
remove the check for a body to the POST in another story.