177 wip 204 no content response likely causing 500 error (#190)
Closes #177 
Closes #189 

Changes the GET /latest to set the status on the response object to 204
instead of returning a JSONResponse with status 204. This was causing
the large exception to appear due to a bug in a library called h11 that
starlette uses.

Updated the GETs that check for objects, like Filings, to return 404s if
those objects don't exist. Uses the same method of setting the
response.status_code vs raising an exception or returning a
Response/JSONResponse.

Updated the ContactInfoDTO to use a RequestValidationException that can
be caught by the exception handlers set in main.py so that the response
error_details will be consistent with our other exceptions. ValueError
is a general python error that wasn't being handle in the same manner.
Could potentially add another exception handler for it in api-commons?
Or just go with the RequestValidationException?