216 load testing with locust (#237)
Closes #216 

Have a couple of shell scripts. One is for ease of use locally, runs it
by building a docker image and then running a headless (with an env MODE
of headless) version. Essentially a one off run that will write a
report.html file. Keeps people from having to install locust locally,
though if someone wants to do that, knock yourself out.

Second script is for the Dockerfile to run, which uses dynamically set
env vars to determine if it's headless vs. single process locust
accessed via the web ui. This could be expanded to include a
master/worker concept if we deem it necessary. Locust can run thousands
of users in a single process which I think is well beyond our needs, but
if we want master/worker, it's easy to do.

The exposing of 5557 is if we want to do the master/worker approach. If
not, I can remove that expose.

I have a conf file that is used for headless and as defaults for the web
ui, but they're overridable there.

I also put in test sblar files that are chosen at random to upload.